Consolidate memory stores into one module and make short-term summarisation opt-in
Collapse the per-type store modules (scope, tokens, models, shortterm, longterm)
into a single kaos_memory/stores.py that holds the scope identity types, token
helpers, ModelClient, ShortTermStore and LongTermStore. This removes a spread of
100-200 line micro-modules in favour of one cohesive storage engine, keeping the
inbound HTTP layer separate while co-locating the outbound model dependency the
stores call.

Unify the store write verb: both ShortTermStore and LongTermStore now expose add()
(mirroring the underlying mem0 .add()), replacing the previous append()/write()
split, and drop the vestigial metadata argument that had no wire representation.

Redesign short-term eviction and summarisation. Rolling summary is now disabled by
default: when off, overflow beyond the token budget and hard event cap is dropped as
a plain recency window with no LLM calls. When explicitly enabled, an overflow marks
the evicted turns and folds the whole chunk into the rolling summary in a single
summariser call off the write path via an optional scheduler, then deletes the folded
rows so storage stays bounded. Guards make DB access thread-safe (lock plus
check_same_thread=False) so the fold can run on a background thread.

Update the tests to the add() verb, the new default, and the batched opt-in folding
and recency-drop behaviours.
