+++ memory-service/tests/test_stores_smoke.py
@@ -12,9 +12,9 @@ from kaos_memory.config import (
     StorageConfig,
     ShortTermTierConfig,
 )
-from kaos_memory.longterm import LongTermStore
-from kaos_memory.scope import Scope, ScopeLevel
-from kaos_memory.shortterm import ShortTermStore
+from kaos_memory.stores import LongTermStore
+from kaos_memory.stores import Scope, ScopeLevel
+from kaos_memory.stores import ShortTermStore
 from tests._fakes import DeterministicEmbedder
 
 OFFLINE = ModelConfig(base_url="http://127.0.0.1:0/v1", model="offline", api_key="t")
@@ -38,15 +38,15 @@ def test_short_term_and_longterm_compose(tmp_path):
     longterm._memory.embedding_model = DeterministicEmbedder()
 
     # Short-term: conversational turns flow into the short-term window.
-    short_term.append(scope, "user", "what port does the deployment use")
-    short_term.append(scope, "assistant", "the deployment uses port 8080")
+    short_term.add(scope, "user", "what port does the deployment use")
+    short_term.add(scope, "assistant", "the deployment uses port 8080")
     assert short_term.recent(scope) == [
         ("user", "what port does the deployment use"),
         ("assistant", "the deployment uses port 8080"),
     ]
 
     # Long-term: a durable fact is written and recalled back under the same scope.
-    longterm.write(scope, "the deployment uses port 8080", infer=False)
+    longterm.add(scope, "the deployment uses port 8080", infer=False)
     hits = longterm.recall(scope, "what port does the deployment use", top_k=5)
     assert any("8080" in h["memory"] for h in hits)
